backend/requirements.txt
fastapi==0.128.0
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
uvicorn==0.40.0
google-auth==2.47.0
google-genai==1.59.0
